requirements.txt
certifi~=2020.4.5.1
chardet
dj-database-url
Django~=2.2
django-ckeditor~=6.0.0
django-froala-editor
django-heroku
django-js-asset
django-recaptcha2
gunicorn
html5lib~=1.0.1
idna
Pillow~=8.2.0
psycopg2-binary
PyPDF2
pytz~=2020.5
reportlab
requests
six~=1.15.0
sqlparse~=0.3.1
urllib3
virtualenv
webencodings~=0.5.1
whitenoise
xhtml2pdf
xlwt

cffi~=1.14.2
PyQt5~=5.15.1
future~=0.18.2
pip
cryptography~=2.8.0
py~=1.9.0
lxml
chatterbot~=1.0.8
pymongo~=3.11.0
nltk~=3.6.1
numpy~=1.19.1
pytest~=6.1.0
click~=7.1.2
tqdm~=4.56.0
regex~=2021.4.4
pyparsing~=2.4.7
Babel~=2.8.0
packaging~=20.4
colorama~=0.4.3
mysqlclient~=2.0.1
asgiref~=3.2.10
Jinja2~=2.11.2
docutils~=0.13.1
joblib~=1.0.1
python-dateutil
setuptools
attrs~=20.2.0
mock~=4.0.2
selenium~=3.141.0
djangorestframework~=3.11.0
uritemplate~=3.0.1
google-api-python-client~=1.8.0
Markdown~=3.2.2
Pygments~=2.6.1
chatbot~=1.5.2b0